requirements.txt
streamlit
moviepy
gTTS
openai-whisper
torch
deep_translator
soundfile
ffmpeg-python
